re-export the structs
- added docs
- added tests
- put struct decls first
- moved errors to separate files
