+++ src/lib.rs
@@ -69,11 +69,17 @@ pub mod stream;
 
 /// Helper functions and types for fixed-length arrays.
 pub mod array {
-    pub use crate::future::race_ok::array::AggregateError;
+    pub use crate::future::join::array::Join;
+    pub use crate::future::race::array::Race;
+    pub use crate::future::race_ok::array::{AggregateError, RaceOk};
+    pub use crate::future::try_join::array::TryJoin;
     pub use crate::stream::merge::array::Merge;
 }
 /// A contiguous growable array type with heap-allocated contents, written `Vec<T>`.
 pub mod vec {
-    pub use crate::future::race_ok::vec::AggregateError;
+    pub use crate::future::join::vec::Join;
+    pub use crate::future::race::vec::Race;
+    pub use crate::future::race_ok::vec::{AggregateError, RaceOk};
+    pub use crate::future::try_join::vec::TryJoin;
     pub use crate::stream::merge::vec::Merge;
 }
